switch to uint64 for node ids (to match with graph code)

+++ runner/brb/brb.go
@@ -10,13 +10,13 @@ type Application interface {
 }
 
 type Network interface {
-	Send(messageType uint8, dest uint16, uid uint32, data []byte)
+	Send(messageType uint8, dest uint64, uid uint32, data []byte)
 }
 
 type Config struct {
 	Byz        bool
-	Id         uint16
-	Neighbours []uint16
+	Id         uint64
+	Neighbours []uint64
 	Graph      *graph.UndirectWeighted
 }
 
@@ -25,7 +25,7 @@ type Protocol interface {
 	// be completed before announcing that they're ready)
 	Init(n Network, app Application, cfg Config)
 
-	Receive(messageType uint8, src uint16, uid uint32, data []byte)
+	Receive(messageType uint8, src uint64, uid uint32, data []byte)
 
 	Send(uid uint32, payload []byte)
 }
